Create RB3 package from RB4 folder
Not fully complete yet, but it'll save some time by automating parts of the conversion.

With "createSongPackage" you can drag in an RB4 song folder and it'll create a songs.dta file for you containing all the original data from the RB4 song

It will check for "forgetool" in your PATH and convert the rbmid to a regular mid, and convert the album art to a png

If forgetool is not in your PATH, it will prompt you for forgetool's location. This info is not saved so if you want to do multiple conversions, please add the forgetool CLI to your PATH

It will grab the converted MIDI and removes the MARKUP track, it will then grab the rbsong venue and merge it. Fog events are included

It then makes a new folder called "RBpkg" and insert the songs.dta, midi file, and mogg file in their places. It will create a lipsync folder with the individual lipsync files. The album art gets placed here too as that you still need to convert manually. It also creates a "gen" folder for you to place the album art and milo in

To do:
-Automate album art conversion
-Automate lipsync compilation to a milo file
-Merge singalong events for songs that have lipsync_ps4 files

Currently both top items require programs that do not have CLI versions on Windows (C3 Con Tools and Onyx). I want to look at potentially making my own procedure to create these files

The bottom one I have some code for already, but I need a break before I adapt it for these singalong events